Added arbitrary camera position
The camera is now defined by a position and a target point. These are arbitrary and can be anywhere in R3. The image plane is now defined relative to the camera parameters instead of explicitly for a fixed point. The method for calculating the pixel location has been changed to support the arbitrary image plane. When moving objects (camera, target, sphere, light source) around the scene, the x-axis appears reversed, except when moving the light source, in which case the other axes seem reversed. Guessing there is an extra/missing (-) sign somewhere but it is too late (12:42 am) to debug.