Success!!
OOP design was NOT failed. I think the issue I was having was not resetting some variables before starting calculations with a new ray. Setting it up for all scene objects to properly inherit variables and functions was also a challenge. Not sure I've ever used a vector of pointers before. I have only used this level of object-orientedness in Java and Python before. I also make use of tuples so I can return both the object and its distance from the same function. This current model is heavily abstracted, supports multiple rays and spheres and arbitrary positions for the camera, light source and all objects. Currently, rays start at the image plane as a design choice but I plan to change that so they all start at the camera to be more realistic in cases where objects intersect with the image plane. Next order of business is to add shadows.