partial shadows
shadows are sometimes half working... 
coordinate frame is still a mystery to me. +x=left, +y=up, +z=out of screen...except for the light source in which case it is reversed, or a least appears to be. diffuse and specular stop working when i try to compute shadows. very weird behavior changes depending on lots of different things. even with only one sphere in the scene it acts like that. who know?? not me for sure. maybe there is something wrong with my function to determine the first intersection. that is the most complex one i have i think. getting real tired of this